Add files via upload
An object of FilterConfig is created by the web container. This object can be used to get the configuration information from the web.xml file.

In this example, if you change the param-value to no, request will be forwarded to the servlet otherwise filter will create the response with the message: this page is under construction.